.history/requirements_20250514190832.txt
pandas
sentence-transformers
faiss-cpu
gpt4all
tkinter
numpy
